+++ sdk/src/configuration/configuration_parser.cc
@@ -256,8 +256,8 @@ ConfigurationParser::ParseOtlpHttpLogRecordExporterConfiguration(
   model->compression  = node->GetString("compression", "");
   model->timeout      = node->GetInteger("timeout", 10000);
 
-  std::string encoding = node->GetString("encoding", "protobuf");
-  model->encoding      = ParseOtlpHttpEncoding(node, encoding);
+  const std::string encoding = node->GetString("encoding", "protobuf");
+  model->encoding            = ParseOtlpHttpEncoding(node, encoding);
 
   return model;
 }
@@ -565,16 +565,17 @@ ConfigurationParser::ParseOtlpHttpPushMetricExporterConfiguration(
   model->compression  = node->GetString("compression", "");
   model->timeout      = node->GetInteger("timeout", 10000);
 
-  std::string temporality_preference = node->GetString("temporality_preference", "cumulative");
-  model->temporality_preference      = ParseTemporalityPreference(node, temporality_preference);
+  const std::string temporality_preference =
+      node->GetString("temporality_preference", "cumulative");
+  model->temporality_preference = ParseTemporalityPreference(node, temporality_preference);
 
-  std::string default_histogram_aggregation =
+  const std::string default_histogram_aggregation =
       node->GetString("default_histogram_aggregation", "explicit_bucket_histogram");
   model->default_histogram_aggregation =
       ParseDefaultHistogramAggregation(node, default_histogram_aggregation);
 
-  std::string encoding = node->GetString("encoding", "protobuf");
-  model->encoding      = ParseOtlpHttpEncoding(node, encoding);
+  const std::string encoding = node->GetString("encoding", "protobuf");
+  model->encoding            = ParseOtlpHttpEncoding(node, encoding);
 
   return model;
 }
@@ -604,10 +605,11 @@ ConfigurationParser::ParseOtlpGrpcPushMetricExporterConfiguration(
   model->compression  = node->GetString("compression", "");
   model->timeout      = node->GetInteger("timeout", 10000);
 
-  std::string temporality_preference = node->GetString("temporality_preference", "cumulative");
-  model->temporality_preference      = ParseTemporalityPreference(node, temporality_preference);
+  const std::string temporality_preference =
+      node->GetString("temporality_preference", "cumulative");
+  model->temporality_preference = ParseTemporalityPreference(node, temporality_preference);
 
-  std::string default_histogram_aggregation =
+  const std::string default_histogram_aggregation =
       node->GetString("default_histogram_aggregation", "explicit_bucket_histogram");
   model->default_histogram_aggregation =
       ParseDefaultHistogramAggregation(node, default_histogram_aggregation);
@@ -624,10 +626,11 @@ ConfigurationParser::ParseOtlpFilePushMetricExporterConfiguration(
 
   model->output_stream = node->GetString("output_stream", "");
 
-  std::string temporality_preference = node->GetString("temporality_preference", "cumulative");
-  model->temporality_preference      = ParseTemporalityPreference(node, temporality_preference);
+  const std::string temporality_preference =
+      node->GetString("temporality_preference", "cumulative");
+  model->temporality_preference = ParseTemporalityPreference(node, temporality_preference);
 
-  std::string default_histogram_aggregation =
+  const std::string default_histogram_aggregation =
       node->GetString("default_histogram_aggregation", "explicit_bucket_histogram");
   model->default_histogram_aggregation =
       ParseDefaultHistogramAggregation(node, default_histogram_aggregation);
@@ -637,11 +640,18 @@ ConfigurationParser::ParseOtlpFilePushMetricExporterConfiguration(
 
 std::unique_ptr<ConsolePushMetricExporterConfiguration>
 ConfigurationParser::ParseConsolePushMetricExporterConfiguration(
-    const std::unique_ptr<DocumentNode> & /* node */) const
+    const std::unique_ptr<DocumentNode> &node) const
 {
   auto model = std::make_unique<ConsolePushMetricExporterConfiguration>();
 
-  // FIXME-CONFIG: https://github.com/open-telemetry/opentelemetry-configuration/issues/242
+  const std::string temporality_preference =
+      node->GetString("temporality_preference", "cumulative");
+  model->temporality_preference = ParseTemporalityPreference(node, temporality_preference);
+
+  const std::string default_histogram_aggregation =
+      node->GetString("default_histogram_aggregation", "explicit_bucket_histogram");
+  model->default_histogram_aggregation =
+      ParseDefaultHistogramAggregation(node, default_histogram_aggregation);
 
   return model;
 }
@@ -1479,8 +1489,8 @@ ConfigurationParser::ParseOtlpHttpSpanExporterConfiguration(
   model->compression  = node->GetString("compression", "");
   model->timeout      = node->GetInteger("timeout", 10000);
 
-  std::string encoding = node->GetString("encoding", "protobuf");
-  model->encoding      = ParseOtlpHttpEncoding(node, encoding);
+  const std::string encoding = node->GetString("encoding", "protobuf");
+  model->encoding            = ParseOtlpHttpEncoding(node, encoding);
 
   return model;
 }